Initilazed the beta stage of music sorter
The main thing that this beta version would do will be described in the
Read me file but for this commit it will also be some elaboration.

This beta version has one main function. The main requirement of this
bot is a super group that has the genres that you'll need as a topic.
for example:
    - I have created a super group with the name of playlist.
    - I have created the genres as a tpoic:
        - Rock
        - Metal
        - Pop
        - Rap
        ....
    - I have found the Tpoic ID via the telegram API or the link
      generated by the link available by copying the link of the
      message. (Read more on the README).
    - After that I would add the topic id to the code.
    - Now just send some music to the bot and caategorize them.