TestingMultipleMessageHandling
Te testing is for the storing of the message ID's of the songs that are being sent to the bot. The bot should be able to save all the ID's and then forward them all to the related topic.

The line that is going to be changed is:

66
context.user_data["audio_message_ids"].append(message.message_id)